refactor: make WebSocketAudioSource a proxy and handle audio decoding in the server itself
